pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amaototyann"
version = "3.0.0"
description = "LINE + Discord notification bot"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "aiohttp>=3.11.0",
    "line-bot-sdk>=3.11.0",
    "discord.py>=2.4.0",
    "pydantic-settings>=2.4.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
]

